unit test fixes and change contcar_to_poscar to be True by default

+++ matmethods/vasp/firetasks/tests/test_copy.py
@@ -10,6 +10,8 @@ __author__ = 'Anubhav Jain <ajain@lbl.gov>'
 module_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)))
 scratch_dir = os.path.join(module_dir, "scratch")
 
+DEBUG = False
+
 class CopyVaspInputsTests(unittest.TestCase):
 
     @classmethod
@@ -23,7 +25,8 @@ class CopyVaspInputsTests(unittest.TestCase):
         os.chdir(scratch_dir)
 
     def tearDown(self):
-        shutil.rmtree(scratch_dir)
+        if not DEBUG:
+            shutil.rmtree(scratch_dir)
 
     def test_unittestsetup(self):
         """
@@ -39,24 +42,30 @@ class CopyVaspInputsTests(unittest.TestCase):
         ct.run_task({})
         files = ["INCAR", "KPOINTS", "POTCAR", "POSCAR"]
         for f in files:
-            self.assertTrue(os.path.exists(os.path.join(module_dir, f)))
+            self.assertTrue(os.path.exists(os.path.join(scratch_dir, f)))
 
         no_files = ["CONTCAR", "OUTCAR"]
         for f in no_files:
-            self.assertFalse(os.path.exists(os.path.join(module_dir, f)))
-
-    def test_plain_copy_more(self):
-        ct = CopyVaspInputs(vasp_dir=self.plain_outdir, contcar_to_poscar=True, additional_files=["OUTCAR"])
-        ct.run_task({})
-        files = ["INCAR", "KPOINTS", "POSCAR", "POTCAR", "OUTCAR"]
-        for f in files:
-            self.assertTrue(os.path.exists(os.path.join(module_dir, f)))
-
-        no_files = ["CONTCAR"]
-        for f in no_files:
-            self.assertFalse(os.path.exists(os.path.join(module_dir, f)))
+            self.assertFalse(os.path.exists(os.path.join(scratch_dir, f)))
 
         # make sure CONTCAR was copied properly
         with open(os.path.join(self.plain_outdir, "CONTCAR")) as f1:
-            with open(os.path.join(module_dir, "POSCAR")) as f2:
+            with open(os.path.join(scratch_dir, "POSCAR")) as f2:
                 self.assertEqual(f1.read(), f2.read())
+
+    def test_plain_copy_more(self):
+        ct = CopyVaspInputs(vasp_dir=self.plain_outdir, contcar_to_poscar=False, additional_files=["OUTCAR"])
+        ct.run_task({})
+        files = ["INCAR", "KPOINTS", "POSCAR", "POTCAR", "OUTCAR"]
+        for f in files:
+            self.assertTrue(os.path.exists(os.path.join(scratch_dir, f)))
+
+        no_files = ["CONTCAR"]
+        for f in no_files:
+            self.assertFalse(os.path.exists(os.path.join(scratch_dir, f)))
+
+        # make sure CONTCAR was NOT copied and POSCAR was instead copied
+        with open(os.path.join(self.plain_outdir, "POSCAR")) as f1:
+            with open(os.path.join(scratch_dir, "POSCAR")) as f2:
+                self.assertEqual(f1.read(), f2.read())
+
